Show dependencies, fix cache and refactor

+++ src/tip/cache.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+import secrets
 
 from tip import config
 
@@ -7,29 +8,36 @@ from tip import config
 CACHE_DIR = config.get('cache_dir')
 
 
-def add(package_dir):
-    """Cache a package at `package_dir` and return path to its cache."""
-    try:
-        os.makedirs(CACHE_DIR, exist_ok=True)
-    except TypeError:
-        raise RuntimeError("Cache is disabled") from None
+def get(package_dir):
+    """Get path to a cached `package_dir` or return it if can't be cached."""
+    if CACHE_DIR is None:
+        return package_dir
     version = os.path.basename(package_dir)
     versions_dir = os.path.dirname(package_dir)
     name = os.path.basename(versions_dir)
     cache_dir = os.path.join(CACHE_DIR, name, version)
     try:
-        shutil.copytree(package_dir, cache_dir)
+        cached_mtime = os.path.getmtime(cache_dir)
+    except FileNotFoundError:
+        cached_mtime = -1
+    if cached_mtime > os.path.getmtime(package_dir):
+        return package_dir
+    os.makedirs(cache_dir, exist_ok=True)
+    temp_dir = os.path.join(CACHE_DIR, name, secrets.token_hex(16) + '~')
+    shutil.copytree(package_dir, temp_dir)
+    try:
+        os.rename(temp_dir, cache_dir)
     except FileExistsError:
         pass
+    finally:
+        shutil.rmtree(temp_dir)
     return cache_dir
 
 
 def clear():
     """Clear cache."""
-    try:
+    if CACHE_DIR is not None and os.path.exists(CACHE_DIR):
         shutil.rmtree(CACHE_DIR)
-    except TypeError:
-        raise RuntimeError("Cache is disabled") from None
 
 
 def is_enabled():
